Only execute the configure ovn-driver role once
The ovirt-provider-ovn-driver role was being executed twice; first
with all default - which was configuring the tunneling network to
'ovirtmgmt' - and then with the desired tunelling network - which,
in a network that does not exist, was not sent.

That caused the playbook to fail, but by then, it was already
executed once.

Change-Id: I4cc27ae2fd1fb5c673409fe486343657cef78736
Bug-Url: https://bugzilla.redhat.com/1613845
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>