build: Automatically adapt for release builds
This patch is meant to reduce the "build" and "post build" patches
needed for each build. Instead of patching the automation scripts,
make them change their behavior based on whether we build for a
milestone or for a release. This is now decided upon purely by checking
if the version in pom.xml ends with 'SNAPSHOT' or not, and is done in
preparation for using maven-release-plugin for builds.

Change-Id: Idf9a50cf5a4d409a38bf2d35b1cbeb29102d515e
Signed-off-by: Yedidyah Bar David <didi@redhat.com>