core: clear migrating_to_vds when vm is up or paused
Previously, we tried to clean the migrating_to_vds field of VMs when
detecting a trigger for that, e.g., when the migration is done or
when the VM goes down. However, there may be cases that complicate
this approach, for instance when the engine restarts during the VM
migration and gets UP after the migration is completed and so the
VM runs on the destination host and there's no trace of the migrate
operation.

Therefore, this patch enhances VMs monitoring with the following
logic: when detecting the VM as running or paused anywhere, the
migrating_to_vds is cleared. That is reasonable as in those cases
the migrate command should have been completed already.

Change-Id: Iaca515408e4a7ae8ff7a579feb903cab3ed6f417
Bug-Url: https://bugzilla.redhat.com/1576134
Signed-off-by: Arik Hadas <ahadas@redhat.com>