webadmin: avoid NPE in case of cloning a non server type snapshot
Fix a NPE caused while trying to clone a snapshot of a non server type
VM.

This NPE was caused due to events chain triggered in webadmin by
VmSnapshotListModel::cloneVM() method while calling the following line:
model.getVmType().setSelectedItem(selectedVm.getVmType());

This started an events chain of updating the vm type which update the
migration mode and that caused the
ExistingVmModelBehavior::updateHaAvailability() method to be called
even before the behavior's vm is set.

This NPE is reproduced on 4.2.6 since only then this specific chain is
triggered, but to be on the safe side I will add
this fix to master as well.

Change-Id: I2860a755ebb5d1236f1b16005af86db7bcd0ac06
Bug-Url: https://bugzilla.redhat.com/1623379
Signed-off-by: Sharon Gratch <sgratch@redhat.com>