Fix header losing focus in Cluster's Manage Networks
Because the header is not part of the table, the focus
stay on the last element that was clicked there. This
caused the focus to be restored at wrong place,
every time the header checkbox was clicked.

To prevent this issue we will steal the focus
and give to the column under the header.

Change-Id: I0002ad8a1268963037019a22b969edae5cec0955
Bug-Url: https://bugzilla.redhat.com/1533389
Signed-off-by: Ales Musil <amusil@redhat.com>