Frontend: allow cloning a snapshot of a headless vm
There was a bug in frontend that prevented from cloning
a snapshot of a headless vm (and also of a high pefromance vm type, which
is always set to headless by default).
The operation was always failed in validation phase of
AddVmFromSnapshotCommand.

The bug is that the headless mode (marked as
getVmStaticData().getDefaultDisplayType() == DisplayType.none) was not
set in the ui model and therefore passed in parameters to the backend command
as a non headless vm instead of headless and that caused the validation
failure.

Change-Id: If8babefcb45aec01e50674fb9c958b82d05cdc3c
Bug-Url: https://bugzilla.redhat.com/1623379
Signed-off-by: Sharon Gratch <sgratch@redhat.com>