restapi: Direct Network Links To API Root
When Network entity is retrieved at the following contexts:

  GET .../api/netowrks
  GET .../api/datacenters/xxx/networks
  GET .../api/iscsibonds/xxx/networks

The href on the network object should direct to the API root:

  .../api/netwoks

And not to the relative context:

  .../api/datacenters/xxx/networks/yyy
  .../api/iscsibonds/xxx/networks/yyy

The reason for the bug is actually a proper repair done to
the repository of static data about the API structure, used
for link-generation purposes. This data was manually maintained
by developers, and thus contained many errors.

This patch:

  https://gerrit.ovirt.org/#/c/82294/

made the repository 100% correct by replacing the human maintenance
with automatic data mining. However in some cases, the result of the
repository containing the correct data broke backwards compatibility,
in places where particular implementations 'counted on' the repository
being wrong, containing a bug (2 wrongs make a right). As such cases
surface, they are handeled manually as in this patch

Bug-Url: https://bugzilla.redhat.com/1638422
Change-Id: Id6e1e4887c59804b25bd6e1bd536be737f0dc6c1
Signed-off-by: Ori_Liel <oliel@redhat.com>