core: allow HA VM with a lease on a non-active storage domain to run
A validation for the VM lease storage domain status was added in order
to solve - https://bugzilla.redhat.com/1561006.
Apparently, this validation damaged the ability of an HA VM with a lease
to start on a different host, if a disaster occurs and causes the SPM
without power management to go down, which will lead to a deactivation
of all the storage domains and the DC.

This patch changes the validation of the VM lease storage domain.
A warning message appears in the log if the VM tries to run while
the VM lease storage domains isn't active instead of failing
the command validation.

Change-Id: Ia459ae6f58dd35e2e44eb0b276dd407cb26d159b
Bug-Url: https://bugzilla.redhat.com/1527249
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>