core: Fix disk's interface null value when cloning a VM
While trying to clone a VM with an attached direct LUN or a shareable
disk, the UI task fails due to a 'VALIDATION_DISK_INTERFACE_NOT_NULL'
warning, while the backend task completed successfully.
Setting a value for the disk's interface deals with the problem above.

Also, direct LUN disks shouldn't be a part of the cloned VM.

Change-Id: I4e93f321d3dc2a0a337cca804ff20401e106cad2
Bug-Url: https://bugzilla.redhat.com/1635942
Signed-off-by: Shani Leviim <sleviim@redhat.com>