core: Fixed Failure When Reimporting VMs
After importing a KVM VM to Ovirt, then removing the VM and re-
importing, the process fails. This is because on success, the vm id
still resides in the VDSM job id list. Failures call the deleteV2VJob
function, which means the same needed to be performed for success.
This patch fixes a previous patch that affected other imports.

Change-Id: Ic1f126531145de85e22840e93ca2871f3166f36e
Bug-Url: https://bugzilla.redhat.com/1625591
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>