core: fix race in VmJobs monitoring
Currently, there is a possible race between the regular
VM monitoring and live merge:

1. A periodic run of the VmsMonitoring starts - Thread A
2. Live merge starts - Thread B
3. Vdsm returns the job id to the engine - Thread B
4. Thread B adds the job to the VmJobsMonitoring repository
5. Thread A checks if the jobs in the repository are still running on
the VM, and since it uses the data from step 1, it is before the
blockjob was added to the VM metadata, and as such it is stale
6. Thread A removes the job prematurely based on the stale data

This is fixed by adding a timestamp to VmJob when when it is added
to the jobs repository. The timestamp will then be compared to
the timestamp when VM stats polling occures (PollVMStatsRefresher#poll).

If the VmJob timestamp is smaller than the fetch time in
PollVMStatsRefresher it means the latest data about the VM is stale and
cannot be used to determine the status of the job, so we will delay
the updating of the jobs repository until the next polling cycle

Change-Id: If0d3b6e933a12877fc1e3fb9fd7d4bdb04165b9c
Bug-Url: https://bugzilla.redhat.com/1585950
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>