core: Add user aliases to memory devices
Until this patch, hot plugged memory devices didn't have user aliases and the
aliases were assigned to them by libvirt.  When libvirt is asked to return a
migratable domain XML, it removes its own aliases and keeps only user aliases.
That means when a VM snapshot is created, the aliases of hot plugged memory
devices are not stored.  Later, when running the VM from the snapshot, either
previewed or committed, Vdsm doesn't find the aliases in the stored domain XML.
When hot unplug of one of those memory devices is requested, Vdsm cannot find
the device and the hot unplug request fails.

There may be several different solutions to this problem.  The simplest
one is to add user aliases to memory devices in Engine.  We do that only when
XML Engine is supported – adding the aliases in the legacy code makes problems
with device identification in Engine without further changes and memory hot
unplug is not supported in cluster version < 4.2 by default anyway.

Change-Id: I35d6df8f2b27e31c8b3a6bdc0e5edbd4fb1cdc4b
Bug-Url: https://bugzilla.redhat.com/1496395
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>