core: Fix default value for VdsLocalDisksLowFreeSpace
Patch [1] tried to decrease default value of VdsLocalDisksLowFreeSpace
to 100 MB, but this patched changed default value only upgrades from
3.4 to 3.5, but all new installations from 3.5 up until now still
contains wrong value 500 MB.

This patch sets the default value for new installations to 100 MB and
also changes the value from 500 MB (wrong default) to 100 MB in existing
installations. But if administrators in existing installations already
changed this value, we will honor their value and don't change it to
default.

[1] https://gerrit.ovirt.org/28308

Change-Id: I9f3285574c2cf499706a6e38d980dd0a4dd041a8
Bug-Url: https://bugzilla.redhat.com/1598447
Signed-off-by: Martin Perina <mperina@redhat.com>