core: Fix UI for moving a disk with a damaged ancestor
Screenshot: https://imgur.com/a/LtKyEEv

In case of trying to move a disk with a damaged ancestor (for example,
the ancestor doesn't have an entry in the 'images' table),
the query 'GetAncestorImagesByImagesIdsQuery' returns a null DiskImage
as the image's ancestor, and causes to a UI error.

Using Collectors.toMap() and receiving that null object as the Map's
ValueMapper function result, causes to a NullPionterException.

Change-Id: I299df85211e51afba3ffeb9ac8cc22de81f6a3ba
Bug-Url: https://bugzilla.redhat.com/1635405
Bug-Url: https://bugzilla.redhat.com/1641326
Signed-off-by: Shani Leviim <sleviim@redhat.com>