engine : Upgrade of hosts causes no warning/error/change of status when running VMs are pinned to hosts
When VMs are pinned to specific hosts
and the admin inisitates an upgrade of
of the host on which is VM is running.
If no other hosts the VM can run on is
available for migration the Upgrade
process should fail and the host needs
to move back to UP status.

Change-Id: I3ea742b509a53813536c6cad2ab64000c02583b0
Bug-Url: https://bugzilla.redhat.com/1631215
Signed-off-by: Ravi Nori <rnori@redhat.com>