engine: Ability to set biosType for a cluster
Added biosType field to cluster and serves as a default for biosType for
all VMs in the cluster.

Screenshot: https://imgur.com/a/pyuNHEd

New value CLUSTER_DEFAULT was added to BiosType enum. If
VmStatic#biosType is set to CLUSTER_DEFAULT, the cluster biosType will
be taken as the effective BIOS Type for the VM.

The setting is available only for clusters of x86 architecture family
that have compatibility version 4.3 or newer.

Change-Id: Ic6ad3ba0208935e8451e584af1415671ed72dd44
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>