engine : Updating template of VM Pool leaves tasks stuck after VMs shutdown
Job is left in state STARTED after VMs from
Pool are shutdown after pool template version
is bumped.

Changes to CoCo and AsyncTaskManager requires
that the parent command needs a callback to end
the job after the child commands have succeeded
when the child command's endAction is handled by
the child command.

Change-Id: If41dc6f11a78d6dd109565fb4621315ca36d7b89
Bug-Url: https://bugzilla.redhat.com/1643826
Signed-off-by: Ravi Nori <rnori@redhat.com>