restapi: Redirect vnic-profiles link
   GET .../ovirt-engine/api/vnicprofiles

Should return vnicprofiles with links to API root:

   /ovirt-engine/api/vnicprofiles/[VNIC-PROFILE-ID]

And not:

  /ovirt-engine/api/networks/[NETWORK-ID]/vnicprofiles/[VNIC-PROFILE-ID]

The problem occurs because of a repair to the repository of static
data about the API structure, used for link-generation purposes.
This data was perviously manually maintained, and thus contained many errors.
Since https://gerrit.ovirt.org/#/c/82294 the repository is maintained
automatically and is 100% correct.

However in some cases, the result of the repository containing the correct
data broke backwards compatibility, in places where particular implementations
'counted on' the repository being wrong. As such cases surface, they are
handeled manually.

Bug-Url: https://bugzilla.redhat.com/1645890
Change-Id: I4e248228dd7918d5d0484c5e474e3391af2a7082
Signed-off-by: Ori_Liel <oliel@redhat.com>