engine: Set GetJobByJobId as user query
Set GetJobByJobId as user query, add OwnerId check to it.

When trying to restore snapshot that includes disks as user
through RestAPI, I got error that 'Cannot revert to Snapshot.
VM's Snapshot does not exist.' This is because of GetJobByJobId
query was allowed only for admins. Thus now GetJobByJobId is user
query and for security reasons GetJobByJobIdQuery also compare
job ownerId and sessionSeqId with current userId and user
sessionSeqId.

Change-Id: I6fb561455625ab3dd6b22319f4b15945063c56f3
Bug-Url: https://bugzilla.redhat.com/1646417
Signed-off-by: Bohdan Iakymets <biakymet@redhat.com>