core: Increase page size for foreman requests
The default page size for Foreman requests is 20.
This might show only partial entries which might not be sufficient for
the user.

The default page size is increased, and exposing configuration value to
let admins change it as needed.

This is not pagination support, but only increasing the size of a single
page that will be returned to the user.

Change-Id: I53e02b1a6717432c9d1e0bfbd79c2e5be815455e
Bug-Url: https://bugzilla.redhat.com/1645007
Signed-off-by: Moti Asayag <masayag@redhat.com>