core: switch back to the uptime reported by vdsm
The calculation of the uptime of the VM on the engine side
represents that of the VM itself (i.e., the QEMU process).
However, users seem to care more about the uptime of the guest
OS (e.g., the uptime should be reset when the guest reboots).

This patch makes the minimal change to revert the change that
moved the uptime calculation to the engine so the uptime we
expose will be the one that VDSM reports which represents the
uptime of the guest OS.

Change-Id: Ida1727b994fa7d08c29f201236df81408542c5bb
Bug-Url: https://bugzilla.redhat.com/1635845
Signed-off-by: Arik Hadas <ahadas@redhat.com>