core: Enforce RSA public keys for SSH connections
By default in apache-sshd-0.12 there are enabled DSA and RSA (in this
order) public keys to work with server public keys. In the past we have
received always RSA public keys (as they are default on EL7), but on
recent Fedora versions sshd were configured with both RSA and ECDSA
public keys, so our sshd client received ECDSA public, which it doesn't
support.

This patch enforces RSA public key only.

Change-Id: I10e6abf77dc74585fa51d51df6bc23b2f652cbaa
Bug-Url: https://bugzilla.redhat.com/1591801
Signed-off-by: Martin Perina <mperina@redhat.com>