upgrade: pki: double check on skip of PKI renewal
Due to recent openssl and gnutls updates in rhel 7.6,
the PKI renewal process is now mandatory because
up-to-date hosts will not accept any more
CA certs without a correct TMZ value in notBefore field.

The CA renewal process can have a relevant impact
on user environment so we are still not forcing it by
default but at least let ask about it twice to be sure that
the user really understand the risk of postponing.

Change-Id: I567881e8d0c3c6bae3facf1dbf19275b22d75225
Bug-Url: https://bugzilla.redhat.com/1649267
Signed-off-by: Simone Tiraboschi <stirabos@redhat.com>
Signed-off-by: Martin Perina <mperina@redhat.com>