engine: Prevent automatic PM on hosts with non-ok external status
Hosts with external-status other than 'OK' should not be automaically
started, stopped or restarted by the way of a power-management operation.

The reasoning behind this is that if some event caused a downgrade in the 
external status of a host, probably something is wrong with the host,
and so to be on the safe side, automatic operations on the host should
be disabled.

This patch takes care of automatic restart of a host via power-management
in case of a netowrk Error which leads to host-non-responsive-treatment,
or in case of hosted-engine (engine running on a host) being down.

Bug-Url: https://bugzilla.redhat.com/1388098
Change-Id: I1f7a982506440d7e1fca3ffa8a84690ec3d38d24
Signed-off-by: Ori_Liel <oliel@redhat.com>