core: Ignore Puppet and Puppet CA Proxies if null
When engine deserializes the host group response from Foreman, it treats
non-existing puppet & puppet CA proxies as '0' due to the primitive
nature of the int.
This leads to foreign key constraint violation when sending a request to
Foreman with '0' id which doesn't exist.

If Puppet proxy id and Puppet CA proxy id aren't associated with host
group, they should be ignored and not sent as part of the request to
foreman.

Change-Id: If6ea397f15bba321b3f631f5d1ed91830a9fc6c0
Bug-Url: https://bugzilla.redhat.com/1506547
Signed-off-by: Moti Asayag <masayag@redhat.com>