engine: Disallow deleting of volume in a Storage Domain
If a volume is present in a storage domain was
deleted, even if the volume was removed but
it existed in webadmin and the engine threw
errors.

Added a validation to check if a volume is
present in a storage domain and prevents
deleting it before removing the corresponding
storage domain.

Change-Id: I4b80050919bec54b3545362ae64d2be09bd85033
Bug-Url: https://bugzilla.redhat.com/1620198
Signed-off-by: Kaustav Majumder <kmajumde@redhat.com>