core: Add security groups support on external networks
Security groups - which are an OpenStack networking API construct -
are applied to the logical ports, and in order to be applied, the
port-security-enabled attribute of the port has to be true.

When a port is created, unless otherwise specified, it will inherit
its port-security-enabled value from the network where it is
created.

This patch adds the port security attribute to the ProviderNetwork
class, and maps it from openstack-java-sdk to business entity, and
from business entity to openstack-java-sdk object.

Bug-Url: https://bugzilla.redhat.com/1539829
Change-Id: I0e9f2060815d35cd1bbd89a4a79df75cd55c1a9d
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>