engine: Introduce CommandCallback finalizeCommand method
When the onFailed or onSucceeded methods of
the callback throw and exception and the method
can no longer be retried as max number of retries
have reached, we need a way to clean up the
command hierarchy and end the job. Sub classes
can override this method to perform the cleanup
logic.

Change-Id: Ib74c49371016cab93584a4404142b7f9418c5b92
Bug-Url: https://bugzilla.redhat.com/1633777
Signed-off-by: Ravi Nori <rnori@redhat.com>