service: Add playbook existence check to AnsibleServlet
The AnsibleServlet uses an async context to run ansible playbooks,
but due to how the AnsibleExecutor is setup, the servlet could not
responds with a failure if the ansible-playbook command fails to
run.

A simple check for the existence of the requested playbook has been
added prior to calling the executor. At least the request will fail
if the playbook does not exist.

Change-Id: I20d2dc26a3265aaba057d989c74756ed1bc05162
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>