engine: Handle EXECUTION_FAILED status in callback
In CommandCallback if the child command
has failed execution the callback currently
waits for the child command's end method to
complete irrespective of whether the child
command's end method is marked for completion
by the parent or not.

In case where the child commands end method
has been marked for completion by the parent
the parent should not wait for the child
command's end method to compelete as it will
never happen. But instead should mark itself
as failed and proceed with the failure logic.

Change-Id: I459ea3e0cddd17ccd6e90bb1ec67917aeb82fef1
Bug-Url: https://bugzilla.redhat.com/1633777
Signed-off-by: Ravi Nori <rnori@redhat.com>