engine: Prevent automatic PM on hosts with non-ok external status - cont'd
Hosts with external-status other than 'OK' should not be automaically
started, stopped or restarted by the way of a power-management operation.

The reasoning behind this is that if some event caused a downgrade in the 
external status of a host, probably something is wrong with the host,
and so to be on the safe side, automatic operations on the host should
be disabled.

This patch takes care of automatic PM operations on a host in the
context of power-saving load balancing.

Bug-Url: https://bugzilla.redhat.com/1388098
Change-Id: I1b005707a30b720cc90969d5062ca21d27012af9
Signed-off-by: Ori_Liel <oliel@redhat.com>