engine: improve memory snapshot registration error msg
When registering a VM with a snapshot that includes memory disks,
The registration will fail in case the memory disks already exist
in the environment on a non-active storage domain.

This patch adds the failed memory snapshot disk name and the name of
the non-active storage domain which holds the memory snapshot disk
to the error message.

Before:
Storage Domain '111', could not be found for memory disk: '222'
in snapshot 's1' (id: '333')

After:
Storage Domain sd1 (id: '111'), could not
be found for memory disk: 'snapshot_memory' (id:
'222') in snapshot 's1' (id: '333')

Change-Id: I9da0fdd0ddda2295cd7fdf2d64587198024ca0eb
Bug-Url: https://bugzilla.redhat.com/1511379
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>