core: fix a bug in manual fencing
When a host with PM defined loose connectivity and the PM is disabled in
the cluster, if the admin starts the host manually from UI or API , a
split brain occurs and HA VMs which were running on the host are
transferred to another host before being released from the problematic
host.
This is a result of a piece of code that tries to restart HA VMs when a
host is started.

The restart of HA VMs after a successful PM start command
does not make sense and was removed.

Change-Id: I6e90652fde5df4cd6bf815633b76e0de29a65848
Bug-Url: https://bugzilla.redhat.com/1641882
Signed-off-by: emesika <emesika@redhat.com>