core: Don't penalize if there are no preferred hosts
PreferredHostsWeightPolicyUnit penalizes hosts that are not among
preferred hosts for a given VM.  However if there are no preferred
hosts for the VM that penalization is useless and, what is worse,
confusing log messages about non-preferred hosts are produced.

Let's produce the log messages and penalties only when there are any
preferred hosts for the given VM.  Let's additionally switch the
logging level used from `log' to `debug' to be consistent with logging
in the other weight policy units.

Change-Id: I86152b80205d9cd8c47cd68aad97c40fb887da86
Bug-Url: https://bugzilla.redhat.com/1650422
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>