restapi: Fix null in disk-attachment link
In some case, the exact flow is not certain, a disk-attachment
may have both a vm-id and a template-id, and then the process
of generating a link for it fails due to ambiguity, and the
result is the string 'null' inside the url, something like:

href="/ovirt-engine/api/vms/69739da6-21a5-4064-a478
      -a223c1763720/null/d75fc53b-6fd3-459f-9c9e-cf6d29beec04"

This patch denotes that when getting a disk attachment in the
context of a VM:

  .../api/vms/xxx/diskattachments

The selected 'parent' will always be VM, even if a template
id exists on the disk-attachemnt object.

The same is done for templates; when getting a disk attachment
in the context of a template:

   .../api/templates/xxx/diskattachemnts

The selected 'parent' will always be Template, even if a VM
id exists on the disk-attachemnt object.

Bug-Url: https://bugzilla.redhat.com/1647018
Change-Id: I1adbd1b66db5981b0552ced48d8859d046e5f155
Signed-off-by: Ori_Liel <oliel@redhat.com>