engine: setup networks commit on success
Instruct vdsm to commit any changes applied during setup
networks immediately upon its successful completion and
if connectivity is successfully re-established with engine.

If setup networks is invoked manually by the user from the
webadmin, the commit on success flag is set according to
the user selection of the 'save network configuration'
checkbox in the setup networks form. In any other flow
where the setup networks command is invoked internally,
the commit on success is set internally by engine to true.

This change is in accordance with corresponding
Change-Id Iea7d0a2a5191462c0d6dc4404052840513a2de9d
on vdsm.
There is no BZ for this patch.

Change-Id: If4c2a000bc0ac2ec9f22aa0e97e8e3d261dfcc25
Signed-off-by: Eitan Raviv <eraviv@redhat.com>