restapi: setup networks commit on success
Instruct vdsm to commit any changes applied during setup networks
immediately upon successful completion of the setup networks process
and if connectivity is successfully re-established with engine.
If this flag is not specified in the request, it is assumed that it was
set to false, which is backward compatible with previous behaviour.

This change is in accordance with corresponding Change-Id
Iea7d0a2a5191462c0d6dc4404052840513a2de9d on vdsm.
There is no BZ for this patch.

Change-Id: I415c0f80a33daa4eb7c81d4b439d434c8eb18e55
Signed-off-by: Eitan Raviv <eraviv@redhat.com>