webadmin: Clean quota list in VM dialog
In the VM create/edit dialog, when a user selects data center that has quota enabled,
and then selects a different one that has quota disabled, the quota list is not
cleared, only hidden, but still contains the old quotas.

This caused problems when creating VM from a template.

This patch clears the list, when a data center without quota is selected.

Change-Id: I0b02934a8536a7cc47e6c11f8a1d87f362ce1344
Bug-Url: https://bugzilla.redhat.com/1648917
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>