core: Default SD for imported Template with lease
When a VM Template with lease is exported, the ID of the storage
domain where the lease is located is saved in the OVF. When this
VM Template is imported, the lease is created by default on the same
storage domain. But, if there is no storage domain with such ID
because it may have been destroyed, the VM Template is not created
at all.

This patch changes this behaviour. If there is no storage domain with
the same Lease ID, the lease is set to the current storage domain
where it is possible.

Change-Id: Ia28ce13a2f1cd0b613904856d54cc4b20446968e
Bug-Url: https://bugzilla.redhat.com/1641703
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>