engine : Upgrade of hosts causes no warning/error/change of status when running VMs are pinned to hosts
HostUpgradeCallback should check for the status
of MaintenanceNumberOfVdssCommand when the status
of the host has been moved back to up after the
migrate has failed or the host stays in up because
the validation of the command failed.

If the MaintenanceNumberOfVdssCommand has failed
the HostUpgradeCommand should be failed and an
audit log message is added indicating that the
host upgrade has failed.

Change-Id: I02c228be5d0b0e3147c29fe6aec8fd58575a3876
Bug-Url: https://bugzilla.redhat.com/1631215
Signed-off-by: Ravi Nori <rnori@redhat.com>