aaa: Log out of VM portal only works at the second try
When retrieving the sso access token
for engine session id the query should
check if the session is valid. A session
could still exist in the backend even after
user logout if there are tasks running for
the session. The session is marked as not
valid which is helpful in determining if the
token should be reused.

Change-Id: If1bbe57c1ea4f74c61e163fe0c0bdc36a29e27aa
Bug-Url: https://bugzilla.redhat.com/1650551
Signed-off-by: Ravi Nori <rnori@redhat.com>