aaa: Remove engine_url reference from Engine SSO
engine_url in authorize endpoint of SSO
has been detected as a security vulerability.
The engine_url is only used in the links on
login and change password page where the
user can click the link to go back to
welcome page.

The engine url for these links can be obtained
from the config and need not be passed using
parameters. So removing all references to
engine_url in Engine SSO side.

Change-Id: I630f1608c4339ce2bec7150f223f67a09fe8f7a8
Bug-Url: https://bugzilla.redhat.com/1656881
Signed-off-by: Ravi Nori <rnori@redhat.com>