aaa: Remove engine_url from login servlets
Cleanup Welcome page, WebAdmin and VmPortal login
servlets to remove engine_url from parameters
passed to oVirt SSO authorize endpoint.

The engine_url which was used to set the link
to welcome page in login screen is not used
and the link is set to ENGINE_URI from engine
config.

Change-Id: Ie891acf0296144394f8612bb83207f81d5d47924
Bug-Url: https://bugzilla.redhat.com/1660925
Signed-off-by: Ravi Nori <rnori@redhat.com>