engine: Allow Adding Host Without Immediate Activation
Allow adding, approving or reinstalling  a host without automatically
activating it afterwards.

This is done in the API by providing activate=false matrix
parameter in the URL of the request:

   POST .../api/hosts/xxx;activate={false,true}         (adding)
   POST .../api/hosts/xxx/approve;activate={false,true} (approving)
   POST .../api/hosts/xxx/install;activate={false,true} (reinstalling)

Or unchecking the existing "Activate Host After Install" checkbox in
"Install Host" dialogue, or the new checkbox by the same title added
to "Add Host" dialogue.

The default bevavior remains activating the host after adding or
approving it, and not activating it after installing it (since
user-initiated re-installation is always done in maintenance mode,
and it's natural for the host to remain in maintenance mode).

This means that in the API if 'activate' is not provided, the default
value assigned to it is 'true' for Add and Approve host operations,
and 'false' for install operation.

Bug-Url: https://bugzilla.redhat.com/1561539
Change-Id: I739fc3bb966772110cac950c554a910c13990a9a
Signed-off-by: Ori_Liel <oliel@redhat.com>