packaging: setup: uninstall: Use sha256
Make setup write sha256 hashes for files it changes instead of md5.

Upgrades from previous versions will still work (and still verify the
md5 hashes if needed).

Do not rewrite existing uninstall files on upgrade. In particular, this
patch is not enough to allow upgrading from earlier versions and then
configuring the machine to use FIPS. It's also not enough for FIPS
support for new setups, but should be enough for that for the uninstall
code.

Change-Id: If2857fd0c02eac4a36a8827ec8a2b7926b54d5e4
Bug-Url: https://bugzilla.redhat.com/1392051
Signed-off-by: Yedidyah Bar David <didi@redhat.com>