packaging: activate port security for >= 4.3 releases
Default port security on an OVN logical port should be enabled
on the 4.3 releases upwards, and disabled on 4.2.

For 4.2 releases, we run on the port_security_enabled default
defined in the ovirt-provider-ovn config file, which is False.

The section featuring the parameter mentioned above can be found
in https://github.com/oVirt/ovirt-provider-ovn#section-network

Change-Id: I3ab753299c39b122d4aaaed6e390b8468cc0990a
Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>