core: remove memory/metadata snapshot disk will remove metadata/memory disk also
Memory and metadata snapshot disks recently exposed to the user.
The user can now perform operations on those disks and remove them.
If one of the memory disks (memory or metadata) was removed the
other disk memory disk is useless.

This patch adds the removal of second snapshot memory disk (memory or metadata)
in case the user requested to remove only one of them.

Change-Id: I7f65f11e2db5ba65dff2cef11ab0727fe3c46471
Bug-Url: https://bugzilla.redhat.com/1648625
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>