restapi: On update VM take custom_compatibility_version from NEXT_RUN
When a running VM is updated through REST API and next_run parameter is
not set to true, the NEXT_RUN snapshot is ignored. In particular, when
compatibility version of the cluster is upgraded, NEXT_RUN snapshot is
be created for every running VM in the cluster and it contains the actual
value of custom_compatibility_version field of the VM.
custom_compatibility_version of the VM itself is changed temporarily to
the compatibility version of the cluster before the upgrade. After VM
shutdown the correct value of custom_compatibility_version will be
restored from the NEXT_RUN snapshot.

But if in the meantime the VM is updated through REST_API and next_run
parameter is not set, the NEXT_RUN snapshot is thrown away and
temporarily changed custom_compatibility_version will not be restored
after VM shutdown.

This patch changes the behaviour of VM update in REST API. Now
custom_compatibility_version is always taken from NEXT_RUN snapshot, if
it exists, regardless of the value of next_run parameter.

Change-Id: Ie0750de0d3e17593f74a5afed6837097b3d3c1e8
Bug-Url: https://bugzilla.redhat.com/1662921
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>