engine: Import OVA Duplicate Disk Image Failure
When Importing a VM OVA Image file that was previously exported from the
system with at least one Disk Image, the Import would fail. This patch
regenerates the Disk and Image IDs when importing as a new entity.

This patch also fixes an issue of only calling the attach image
function only when the image details are successfully added to
the Database to avoid further failures.

Change-Id: I6ea437b39dd46fb2677998c49660ff83d11a2dee
Bug-Url: https://bugzilla.redhat.com/1656092
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>