packaging: setup: uninstall: Do not fail on missing files
If a file was written by a run of engine-setup, and removed by a later
run, another later run will find it mentioned in uninstall, but will not
set 'changed' for it.

Do not fail in this case.

I considered changing the code setting 'changed', but could not decide
to what - True is technically correct, but will needlessly list the file
in most cases, False also makes sense but can make us ignore cases we
shouldn't have. Best is probably to change it to a more-than-two-values
field, but I didn't bother for now.

Specific example: on upgrade between pg versions, we move
postgresql.conf between the data dirs. So the former one, written by a
previous engine-setup, will be missing later.

Change-Id: I2264757637a918f1b1de57edde966ea4e3f8ce80
Signed-off-by: Yedidyah Bar David <didi@redhat.com>