core: Allow display network on IPv6
There is no reason to enforce IPv4 for the display network,
but on the contrary, this would even contradict our preferred
IPv6 only configuration.
For this reason, it is allowed to schedule a VM on a host with
an interface on display network, which has an IPv4 or IPv6
boot protocol.

Change-Id: I1d7edd06e92fa9f025971176466b97206e6e55eb
Signed-off-by: Dominik Holler <dholler@redhat.com>