restapi: Unsupport V3 of API from version 4.3 of Engine
Version 3 of the API is not longer suppored as of version 4.3
of the engine. This does not mean that it's unusable, but rather
that fixes and adjustments aren't made anymore to it and users
should be aware of that. Using version 3 of the API will result
in the following audit warning message:

Client ... is using version 3 of the API, which has been deprecated
since version 4.0 of the engine, and will no longer be supported
starting with version 4.3. It is highly recommended to update the
client to use a supported version of the API and the SDKs before
upgrading to version 4.3 of the engine.

Bug-Url: https://bugzilla.redhat.com/1399750
Change-Id: Ia4325f33780470be46c69b76ac7fba8e26e5c2b1
Signed-off-by: Ori_Liel <oliel@redhat.com>