core: validate VM lease exists on storage before adding it
When importing a VM from a configuration/export domain,
the engine automatically creates a new VM lease for a
VM which set as HA VM with a lease.

In a DR scenario, the VM lease should already exist on
the storage domain since all the data should be backed up.
Therefore, we should avoid creating a new lease which
will take time and eventually, will fail the initial
running of the HA VM in the failover/failback script.

Moreover, adding the inspection whether the lease already exists,
will save the need to automatically create it which eventually will
save time when importing a VM to the data-center

Change-Id: Id73318fdd13aaf29fe09a7b6c3c32c761d0e62e6
Bug-Url: https://bugzilla.redhat.com/1676427
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>