webadmin: expose ipv6 gateway on non-mgmt networks
The ipv6 default routes and gateway handling differ from those of ipv4
because in ipv6 source routing is not implemented.

The ipv6 default route of a host is determined from the ipv6 gateways on
its NICs. To support the configuration of the ipv6 default route on any
network of the host from engine, it is therefore necessary to expose the
ipv6 gateway on non-management networks in the setup networks facility.
Until now it was exposed only on the management network.

Exposing it in setup-networks on non-management network attachments means
that it is displayed for all configuration types, and is editable in a
static configuration.

The behavior of the ipv6 gateway on non-management network attachments
is similar to that of the management network:
- it is passed in the setup networks to vdsm (for static configuration)
- it is added to the ifcfg of the relevant NIC by vdsm (for static
  configuration)
- it is reported back in getCaps (only if a single one is define)
- it is saved in engine both on the vds_interface and on the network
  attachment objects and db tables.

A consequence of this patch is that more than one statically configured
ipv6 gateway will be added as ipv6 default routes to the routing table on
the host. This is in addition to the initial default route created from
the management network.

In this patch there is no association between the ipv6 default routes on
the host and the default route role network in engine. The behaviour of
the default route role continues to dependent on ipv4 only.

Change-Id: I5aec5b89a59b52f2fd077501eb89272587ba9b6c
Bug-Url: https://bugzilla.redhat.com/1467332
Signed-off-by: Eitan Raviv <eraviv@redhat.com>