core: equate ipv6 gateway "::" with "no gateway"
When no gateway is specified in a static ipv6 configuration of a
network attachment, it is designated in the network attachment as
null (see change id Ic626a230c193e7aa962c089e2d63847934ed2d18).
This 'no gateway' configuration is passed in setup networks to vdsm
which interprets it as the "::" gateway and returns "::" in getCaps
to engine. To avoid out-of-sync reported because of these different
'no gateway' designations this patch adds a specific check for this
situation within the sync check.

Change-Id: Ia472162a59cef70aee40c1a02e231b6b7c0ef1b0
Bug-Url: https://bugzilla.redhat.com/1607118
Signed-off-by: Eitan Raviv <eraviv@redhat.com>