service: engine proxy handling for AnsibleServlet
In an environment where the engine proxy is enabled, the engine ports are
configured as "None". The AnsibleServlet would fail trying to parse "None"
as an integer. When the engine proxy is enabled, the proxy port is used
instead.

Also adjusted the playbook detection block so the servlet will still
output a log file when the requested playbook doesn't exist.

Change-Id: I92ade879de2f579f3c2424f1bf7d0792c8e8867d
Bug-Url: https://bugzilla.redhat.com/1637593
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>